requirements.txt
# crewai==0.28.8
# crewai[tools]==0.28.8
# python-dotenv==1.0.0
# #openai==1.12.0
# langchain >=0.1.10 , <0.2.0
# openai >=1.13.3 , <2.0.0

#langchain==0.1.0
#langchain-openai==0.0.2
#imapclient==2.3.1
email-validator==2.1.0

crewai==0.28.8
python-dotenv==1.0.0
openai>=1.13.3,<2.0.0
langchain>=0.1.10,<0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
imapclient == 2.3.1
onnxruntime>=1.16.0